requirements.txt
# Azure Functions
azure-functions==1.18.0

# API Clients
requests==2.31.0
openai==1.58.1

# Timezone handling
pytz==2024.1

# Azure Key Vault (for production)
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
